Fix for racecondition on startup with target architecture fix.
The target architecture fix introduced a race condition between
the -exec-run or -target-attach and the module loads finishing.
This was causing many debuggees to fail during attach or startup
due to the process state not being able to send the console command
to obtain the architecture. This moves the acquisition of that state
to between the -file-exec-and-symbols and the actual running.

NOTE: needed to set the binary on attach so that the debuggers
know the architecture. Apparently, the miengine wasn't doing that
before.

CR feedback

1) Honor launch options target architecture if they are there.
NOTE: custom setup commands will throw if it isn't there
2) Remove switching escape to protected from private.

NOTE: After Paul's change, the launch options still contain
the targetarchitecture even though they aren't getting used.
we should make a checkin to remove them for the c++ extension.